+++ packages/utils/test/schema/getDefaultFormStateTest.ts
@@ -1,5 +1,6 @@
 import { createSchemaUtils, getDefaultFormState, RJSFSchema } from '../../src';
 import { computeDefaults } from '../../src/schema/getDefaultFormState';
+import { RECURSIVE_REF, RECURSIVE_REF_ALLOF } from '../testUtils/testData';
 import { TestValidatorType } from './types';
 
 export default function getDefaultFormStateTest(testValidator: TestValidatorType) {
@@ -181,6 +182,16 @@ export default function getDefaultFormStateTest(testValidator: TestValidatorType
         expect(computeDefaults(testValidator, schema, undefined, schema, undefined, 'excludeObjectChildren')).toEqual(
           {}
         );
+      });
+      it('test with a recursive schema', () => {
+        expect(computeDefaults(testValidator, RECURSIVE_REF, undefined, RECURSIVE_REF)).toEqual({
+          name: '',
+        });
+      });
+      it('test with a recursive allof schema', () => {
+        expect(computeDefaults(testValidator, RECURSIVE_REF_ALLOF, undefined, RECURSIVE_REF_ALLOF)).toEqual({
+          value: [undefined],
+        });
       });
     });
     describe('root default', () => {